requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
azure-ai-textanalytics==5.3.0b1
azure-common==1.1.28
azure-core==1.26.3
azure-identity==1.12.0
blis==0.7.11
catalogue==2.0.10
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
cryptography==39.0.1
cymem==2.0.8
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.1
frozenlist==1.3.3
ginza==5.2.0
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
isodate==0.6.1
ja-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ja_core_news_sm-3.7.0/ja_core_news_sm-3.7.0-py3-none-any.whl#sha256=1191e5bbffcc90670146616c274a64850e54d12070bc5846e78a094f2f6fcfca
ja-ginza==5.2.0
Jinja2==3.1.4
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mecab-python3==1.0.9
msal==1.21.0
msal-extensions==1.0.0
multidict==6.0.4
murmurhash==1.0.10
numpy==1.26.4
openai==0.27.2
packaging==24.1
plac==1.4.3
portalocker==2.7.0
preshed==3.0.9
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.6.0
pyzmq==25.0.0
requests==2.28.2
rich==13.7.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-pkuseg==0.0.33
srsly==2.4.8
SudachiDict-core==20240409
SudachiPy==0.6.8
thinc==8.2.4
thinc_apple_ops==0.1.5
tomli==2.0.1
tqdm==4.65.0
typer==0.12.3
typing_extensions==4.12.2
unidic-lite==1.0.8
urllib3==1.26.14
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
yarl==1.8.2
zh-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.7.0/zh_core_web_sm-3.7.0-py3-none-any.whl#sha256=f51075665749e07406d629d1055ce5a68635fae6ab3c34257ee798c62b4fc431
